added service for Opgave
