+++ vue/src/apiService.js
@@ -2,20 +2,8 @@ import axios from 'axios';
 
 const API_URL = 'http://127.0.0.1:8080/api';
 
-export const createOpgave = (data) => {
-  return axios.post(`${API_URL}/opgave`, data);
-};
-
 export const createForloeb = (data) => {
   return axios.post(`${API_URL}/forloeb`, data);
-};
-
-export const getOpgaverByForloebID = (forloebID) => {
-  return axios.get(`${API_URL}/opgave/forloeb/${forloebID}`);
-};
-
-export const getOpgaverByForloebsskabelonID = (forloebsskabelonID) => {
-  return axios.get(`${API_URL}/opgave/forloebsskabelon/${forloebsskabelonID}`);
 };
 
 export const createRessource = (data) => {
@@ -34,10 +22,6 @@ export const getForloebsskabeloner = () => {
   return axios.get(`${API_URL}/forlobsskabelon`);
 };
 
-export const getOpgaver = () => {
-  return axios.get(`${API_URL}/opgave`);
-};
-
 export const getOpgaveskabeloner = () => {
   return axios.get(`${API_URL}/opgaveskabelon`);
 };
@@ -45,10 +29,6 @@ export const getOpgaveskabeloner = () => {
 export const getForloebsskabelonerWithOpgavers = () => {
   return axios.get(`${API_URL}/forlobsskabelon/opgaver`);
 };
-
-export const deleteOpgave = (opgaveID) => {
-  return axios.delete(`${API_URL}/opgave/${opgaveID}`);
-}
 
 export const updateOpgave = (opgaveID, data) => {
   return axios.put(`${API_URL}/opgave/${opgaveID}`, data);
